backend/user-service/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-environ==0.11.2
django-filter==24.3
django-oauth-toolkit==3.0.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
drf-social-oauth2==3.1.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.7
oauthlib==3.2.2
packaging==24.2
phonenumberslite==8.13.52
pika==1.3.2
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
